Add files via upload
Initial upload. In this notebook I provide three models that can be used to calculate a call or put stock option price: the Binomial model, the Black Scholes model, and the Monte Carlo model. I also briefly introduce the QuantLib python package, which can be used to calculate an option price.